+++ be/src/olap/rowset/segment_v2/inverted_index_reader.h
@@ -80,7 +80,7 @@ public:
         io::Path io_path(_path);
         _index_dir = io_path.parent_path();
         _index_file_name = InvertedIndexDescriptor::get_index_file_name(
-                io_path.filename(), index_meta->index_id(), index_meta->get_index_suffix());
+                io_path.filename(), (uint32_t)index_meta->index_id(), index_meta->get_index_suffix());
     }
     virtual ~InvertedIndexReader() = default;
 
@@ -99,7 +99,7 @@ public:
 
     virtual InvertedIndexReaderType type() = 0;
 
-    [[nodiscard]] uint32_t get_index_id() const { return _index_meta.index_id(); }
+    [[nodiscard]] uint32_t get_index_id() const { return (uint32_t)_index_meta.index_id(); }
 
     [[nodiscard]] const std::map<string, string>& get_index_properties() const {
         return _index_meta.properties();
